chore: optimize Nuxt build config
& Bootstrap Vue plugin registration
