+++ nuxt.config.js
@@ -1,5 +1,10 @@
 /* eslint-disable camelcase */
 
+/*
+** Nuxt config
+** Docs: https://nuxtjs.org/docs/configuration-glossary/
+*/
+
 const APP_SITE_NAME = 'Europeana';
 
 import pkg from './package.json';
@@ -208,10 +213,16 @@ export default {
     bootstrapVueCSS: false,
 
     // Tree shake plugins
+    //
+    // NOTE: do not register plugins globally (here) unless they are used widely;
+    //       import them locally into the views that need them instead. This
+    //       is to prevent large amounts of unused JS being sent upfront to clients.
+    //       As a general rule, only register globally if at least three views
+    //       use the plugin's components/directives. Also consider how often
+    //       those components are rendered based on placement in layout and
+    //       usage patterns by users, and the plugin's bundled size.
     componentPlugins: [
-      'AlertPlugin',
       'BadgePlugin',
-      'BreadcrumbPlugin',
       'ButtonPlugin',
       'CardPlugin',
       'DropdownPlugin',
@@ -227,16 +238,10 @@ export default {
       'LayoutPlugin',
       'LinkPlugin',
       'ListGroupPlugin',
-      'MediaPlugin',
       'ModalPlugin',
       'NavbarPlugin',
-      'NavPlugin',
-      'PaginationNavPlugin',
       'SidebarPlugin',
-      'TablePlugin',
-      'TabsPlugin',
-      'ToastPlugin',
-      'TooltipPlugin'
+      'ToastPlugin'
     ]
   },
 
@@ -362,7 +367,26 @@ export default {
   /*
   ** Build configuration
   */
-  build: {},
+  build: {
+    // Extract CSS to files instead of inlining to keep HTML page size down.
+    // This results in more network requests in total to render the full page,
+    // but a smaller HTML document, and a smaller total size of all requests.
+    // Lighthouse performance ratings favour this by a few percentage points.
+    extractCSS: true,
+
+    extend(config, { isClient }) {
+      // Extend webpack config only for client bundle
+      if (isClient) {
+        // Build source maps to aid debugging in production builds
+        config.devtool = 'source-map';
+      }
+    }
+  },
+
+  /*
+  ** Enable modern builds
+  */
+  modern: true,
 
   /*
   ** Render configuration
