+++ src/components/generic/AlertMessage.vue
@@ -10,7 +10,12 @@
 </template>
 
 <script>
+  import { BAlert } from 'bootstrap-vue';
+
   export default {
+    components: {
+      BAlert
+    },
     props: {
       error: {
         type: String,
