+++ src/components/generic/InfoMessage.vue
@@ -20,7 +20,12 @@
 </template>
 
 <script>
+  import { BAlert } from 'bootstrap-vue';
+
   export default {
+    components: {
+      BAlert
+    },
     props: {
       message: {
         type: String,
